refactor: Move to entity module
