+++ lib/src/internal/managers/module_manager.dart
@@ -8,26 +8,6 @@ import 'package:mineral/src/internal/managers/store_manager.dart';
 
 import 'event_manager.dart';
 
-class Module {
-  final String identifier;
-  final String label;
-  final String? description;
-
-  const Module({ required this.identifier, required this.label, this.description });
-}
-
-abstract class MineralModule {
-  late final String identifier;
-  late final String label;
-  late final String? description;
-
-  abstract List<MineralEvent> events;
-  abstract List<MineralCommand> commands;
-  abstract List<MineralStore> stores;
-
-  Future<void> init () async {}
-}
-
 class ModuleManager {
   final Map<String, MineralModule> _modules = {};
   Map<String, MineralModule> get modules => _modules;
